requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dateutil
feedgen
email-validator
slowapi
sqlalchemy
passlib
bcrypt
resend
python-dotenv
openai
requests
feedparser
beautifulsoup4
newspaper3k
lxml_html_clean
azure-storage-blob
